可配置的日志时间格式

+++ src/util/DateTime.cpp
@@ -15,7 +15,7 @@ int gettimeofday(struct timeval *tp, void *tzp) {
     tm.tm_sec = wtm.wSecond;
     tm.tm_isdst = -1;
     clock = mktime(&tm);
-    tp->tv_sec = clock;
+    tp->tv_sec = (long) clock;
     tp->tv_usec = wtm.wMilliseconds * 1000;
     return 0;
 }
